move code around so that differnt evolve methods are in different source files
