+++ setup.py
@@ -74,7 +74,9 @@ ext_modules = [
         Extension(
             'fwdpy11_arg_example.wfarg',
             ['fwdpy11_arg_example/wfarg.cc',
-                # 'fwdpy11_arg_example/ancestry_edge_sets.cc',
+                'fwdpy11_arg_example/wright_fisher_single_thread.cc',
+                'fwdpy11_arg_example/wright_fisher_async.cc',
+                'fwdpy11_arg_example/wright_fisher_queue.cc',
                 'fwdpy11_arg_example/handle_recombination.cc'
                 ],
             depends = glob.glob('fwdpy11_arg_example/*.hpp'),
